+++ src/main/java/org/sosly/witchcraft/rituals/RitualRegistry.java
@@ -7,6 +7,7 @@ import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.registries.RegisterEvent;
 import org.sosly.witchcraft.Witchcraft;
 import org.sosly.witchcraft.rituals.effects.BrokenSympathyRitual;
+import org.sosly.witchcraft.rituals.effects.HedgeRitual;
 import org.sosly.witchcraft.rituals.effects.SympathyRitual;
 
 @Mod.EventBusSubscriber(modid = Witchcraft.MOD_ID, bus = Mod.EventBusSubscriber.Bus.MOD)
@@ -15,6 +16,7 @@ public class RitualRegistry {
     public static void registerRitualEffects(RegisterEvent event) {
         event.register(Registries.RitualEffect.get().getRegistryKey(), helper -> {
             helper.register("rituals.broken_sympathy", new BrokenSympathyRitual(new ResourceLocation(Witchcraft.MOD_ID, "rituals/broken_sympathy")));
+            helper.register("rituals.hedge", new HedgeRitual(new ResourceLocation(Witchcraft.MOD_ID, "rituals/hedge")));
             helper.register("rituals.sympathy", new SympathyRitual(new ResourceLocation(Witchcraft.MOD_ID, "rituals/sympathy")));
         });
     }